WIP: Upgrading test app to Rails 5
I'm working on getting `rake test` to pass. The Rails app in
`test/dummy` is a Rails 4 app, and it seems as though it needs to be
upgraded in order to get the tests passing.

I'm currently getting an error about `Rack::Lock` not being available. I
did try using an older version of Bundler (1.17.3) and an older version
of Rails (2.5.3), but I get the same error.